pyproject.toml
[project]
name = "02-deep-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pillow>=12.1.1",
    "tensorboard>=2.20.0",
    "torch",
    "torchvision",
    "torchsummary>=1.5.1",
    "setuptools==80.10.2",
    "tqdm>=4.67.3",
    "bing-image-downloader>=1.1.2",
    "koreanize-matplotlib>=0.1.1",
    "grad-cam>=1.5.5",
    "fastapi>=0.133.0",
    "uvicorn>=0.41.0",
    "uuid>=1.30",
    "python-multipart>=0.0.22",
    "streamlit>=1.54.0",
    "plotly>=6.6.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
